# swagger_json.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/cache/warmup/user/{userId}": {
      "post": {
        "operationId": "CacheMonitorController_warmupUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "用户缓存预热完成"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "预热指定用户的缓存",
        "tags": [
          "cache"
        ]
      }
    },
...
    "/api/users": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "用户创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "409": {
            "description": "用户已存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建用户",
        "tags": [
          "用户管理"
        ]
      },
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "搜索关键词",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleId",
            "required": false,
            "in": "query",
            "description": "用户角色ID筛选",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "minimum": 10,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "排序字段",
            "schema": {
              "default": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "排序方向",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取用户列表成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取用户列表",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/api/users/search/by-email": {
      "get": {
        "operationId": "UsersController_searchByEmail",
        "parameters": [
          {
            "name": "email",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "搜索成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSearchResultDto"
                }
              }
            }
          },
          "404": {
            "description": "用户不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "根据邮箱搜索用户",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/api/users/search": {
      "get": {
        "operationId": "UsersController_searchUsers",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "搜索关键词",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleId",
            "required": false,
            "in": "query",
            "description": "用户角色ID筛选",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "minimum": 10,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "排序字段",
            "schema": {
              "default": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "排序方向",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "搜索成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "搜索用户（用于添加项目成员）",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/api/users/profile/me": {
      "get": {
        "operationId": "UsersController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "获取用户信息成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取当前用户信息",
        "tags": [
          "用户管理"
        ]
      },
      "patch": {
        "operationId": "UsersController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新用户信息成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新当前用户信息",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取用户成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "用户不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "根据 ID 获取用户",
        "tags": [
          "用户管理"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新用户成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "用户不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新用户",
        "tags": [
          "用户管理"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除用户成功"
          },
          "404": {
            "description": "用户不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "删除用户",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/api/users/{id}/status": {
      "patch": {
        "operationId": "UsersController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "更新用户状态成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "用户不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新用户状态",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/api/users/change-password": {
      "post": {
        "operationId": "UsersController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "密码修改成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangePasswordApiResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权或旧密码不正确"
          },
          "409": {
            "description": "旧密码不正确"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "修改密码",
        "tags": [
          "用户管理"
        ]
      }
    },
...
    "/api/admin/permissions/cache/user/{userId}": {
      "delete": {
        "operationId": "AdminController_clearUserCache",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "用户权限缓存已清除",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCacheClearResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "清除用户权限缓存",
        "tags": [
          "管理员"
        ]
      }
    },
    "/api/admin/permissions/user/{userId}": {
      "get": {
        "operationId": "AdminController_getUserPermissions",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取用户权限信息成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPermissionsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取用户权限信息",
        "tags": [
          "管理员"
        ]
      }
    },
...
  },
  "info": {
    "title": "CloudCAD API",
    "description": "图纸管理平台API文档",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UserRoleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "角色 ID"
          },
          "name": {
            "type": "string",
            "description": "角色名称"
          },
          "description": {
            "type": "string",
            "description": "角色描述"
          },
          "isSystem": {
            "type": "boolean",
            "description": "是否为系统角色"
          }
        },
        "required": [
          "id",
          "name",
          "isSystem"
        ]
      },
      "UserProfileResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "用户 ID"
          },
          "email": {
            "type": "string",
            "description": "用户邮箱"
          },
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "nickname": {
            "type": "string",
            "description": "用户昵称"
          },
          "avatar": {
            "type": "string",
            "description": "头像 URL"
          },
          "status": {
            "type": "string",
            "description": "用户状态",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "SUSPENDED"
            ]
          },
          "role": {
            "description": "用户角色",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRoleDto"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "创建时间"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "更新时间"
          }
        },
        "required": [
          "id",
          "email",
          "username",
          "status",
          "role",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "用户邮箱",
            "example": "user@example.com",
            "format": "email"
          },
          "username": {
            "type": "string",
            "description": "用户名",
            "example": "username",
            "minLength": 3,
            "maxLength": 20,
            "pattern": "^[a-zA-Z0-9_]+$"
          },
          "password": {
            "type": "string",
            "description": "密码",
            "example": "password123",
            "minLength": 8,
            "maxLength": 50
          },
          "nickname": {
            "type": "string",
            "description": "昵称",
            "maxLength": 50
          },
          "avatar": {
            "type": "string",
            "description": "头像URL"
          },
          "roleId": {
            "type": "string",
            "description": "角色ID",
            "example": "clh8x9y0z1a2b3c4d5e6f7g8h9"
          }
        },
        "required": [
          "email",
          "username",
          "password"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "用户 ID"
          },
          "email": {
            "type": "string",
            "description": "用户邮箱"
          },
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "nickname": {
            "type": "string",
            "description": "用户昵称"
          },
          "avatar": {
            "type": "string",
            "description": "头像 URL"
          },
          "status": {
            "type": "string",
            "description": "用户状态",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "SUSPENDED"
            ]
          },
          "role": {
            "description": "用户角色",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRoleDto"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "创建时间"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "更新时间"
          }
        },
        "required": [
          "id",
          "email",
          "username",
          "status",
          "role",
          "createdAt",
          "updatedAt"
        ]
      },
      "UserListResponseDto": {
        "type": "object",
        "properties": {
          "users": {
            "description": "用户列表",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserResponseDto"
            }
          },
          "total": {
            "type": "number",
            "description": "总数"
          },
          "page": {
            "type": "number",
            "description": "当前页码"
          },
          "limit": {
            "type": "number",
            "description": "每页数量"
          },
          "totalPages": {
            "type": "number",
            "description": "总页数"
          }
        },
        "required": [
          "users",
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
      "UserSearchResultDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "用户 ID"
          },
          "email": {
            "type": "string",
            "description": "用户邮箱"
          },
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "nickname": {
            "type": "string",
            "description": "用户昵称"
          },
          "avatar": {
            "type": "string",
            "description": "头像 URL"
          }
        },
        "required": [
          "id",
          "email",
          "username"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "用户邮箱"
          },
          "username": {
            "type": "string",
            "description": "用户名",
            "minLength": 3
          },
          "password": {
            "type": "string",
            "description": "密码",
            "minLength": 6
          },
          "nickname": {
            "type": "string",
            "description": "昵称"
          },
          "avatar": {
            "type": "string",
            "description": "头像URL"
          },
          "roleId": {
            "type": "string",
            "description": "角色ID"
          },
          "status": {
            "type": "string",
            "description": "用户状态",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "SUSPENDED"
            ]
          }
        }
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "oldPassword": {
            "type": "string",
            "description": "旧密码",
            "example": "OldPassword123!"
          },
          "newPassword": {
            "type": "string",
            "description": "新密码",
            "example": "NewPassword123!"
          }
        },
        "required": [
          "oldPassword",
          "newPassword"
        ]
      },
      "ChangePasswordResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "消息"
          }
        },
        "required": [
          "message"
        ]
      },
      "ChangePasswordApiResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "响应状态码",
            "example": "SUCCESS",
            "enum": [
              "SUCCESS",
              "ERROR"
            ]
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          },
          "data": {
            "description": "响应数据",
            "allOf": [
              {
                "$ref": "#/components/schemas/ChangePasswordResponseDto"
              }
            ]
          },
          "timestamp": {
            "type": "string",
            "description": "响应时间戳",
            "example": "2025-12-12T03:34:55.801Z"
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "timestamp"
        ]
      },
...
      "UserCacheClearResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "提示消息"
          }
        },
        "required": [
          "message"
        ]
      },
      "UserPermissionInfoDto": {
        "type": "object",
        "properties": {
          "userRole": {
            "type": "string",
            "description": "用户角色"
          },
          "permissions": {
            "description": "权限列表",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "userRole",
          "permissions"
        ]
      },
      "UserPermissionsResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "提示消息"
          },
          "data": {
            "description": "用户权限信息",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserPermissionInfoDto"
              }
            ]
          }
        },
        "required": [
          "message",
          "data"
        ]
      },
...
    }
  }
}